# backend/openapi.json
{
  "components": {
    "schemas": {
      "Category": {
        "enum": [
          "water",
          "electricity",
          "sanitation",
          "roads",
          "streetlights",
          "other"
        ],
        "title": "Category",
        "type": "string"
      },
      "ComplaintCreate": {
        "properties": {
          "location": {
            "maxLength": 200,
            "minLength": 3,
            "title": "Location",
            "type": "string"
          },
          "reporter_contact": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Contact"
          },
          "text": {
            "maxLength": 2000,
            "minLength": 10,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "location"
        ],
        "title": "ComplaintCreate",
        "type": "object"
      },
      "ComplaintListRead": {
        "description": "The paginated complaint collection returned by ``GET /api/complaints``.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ComplaintRead"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "ComplaintListRead",
        "type": "object"
      },
      "ComplaintRead": {
        "properties": {
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "reporter_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter Contact"
          },
          "status": {
            "$ref": "#/components/schemas/ComplaintStatus"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "triage_latency_ms": {
            "title": "Triage Latency Ms",
            "type": "integer"
          },
          "triaged_by": {
            "title": "Triaged By",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "location",
          "reporter_contact",
          "category",
          "priority",
          "status",
          "ai_summary",
          "triaged_by",
          "triage_latency_ms"
        ],
        "title": "ComplaintRead",
        "type": "object"
      },
      "ComplaintStatsRead": {
        "description": "Aggregate counts used by the operations dashboard.",
        "properties": {
          "by_category": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Category"
            },
            "title": "By Category",
            "type": "object"
          },
          "by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/ComplaintStatus"
            },
            "title": "By Status",
            "type": "object"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "by_status",
          "by_category"
        ],
        "title": "ComplaintStatsRead",
        "type": "object"
      },
      "ComplaintStatus": {
        "enum": [
          "open",
          "in_progress",
          "resolved",
          "rejected"
        ],
        "title": "ComplaintStatus",
        "type": "string"
      },
      "ComplaintStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ComplaintStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "ComplaintStatusUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Priority": {
        "enum": [
          "high",
          "normal",
          "low"
        ],
        "title": "Priority",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "CivicPulse API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/complaints": {
      "get": {
        "operationId": "list_complaints_api_complaints_get",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Category"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Priority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ComplaintStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintListRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Complaints",
        "tags": [
          "complaints"
        ]
      },
      "post": {
        "operationId": "create_complaint_api_complaints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Complaint",
        "tags": [
          "complaints"
        ]
      }
    },
    "/api/complaints/{complaint_id}": {
      "get": {
        "operationId": "get_complaint_api_complaints__complaint_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "complaint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Complaint Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Complaint",
        "tags": [
          "complaints"
        ]
      }
    },
    "/api/complaints/{complaint_id}/status": {
      "patch": {
        "operationId": "update_complaint_status_api_complaints__complaint_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "complaint_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Complaint Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Complaint Status",
        "tags": [
          "complaints"
        ]
      }
    },
    "/api/meta/providers": {
      "get": {
        "operationId": "provider_metadata_api_meta_providers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Provider Metadata Api Meta Providers Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Provider Metadata",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/stats": {
      "get": {
        "operationId": "statistics_api_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplaintStatsRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Statistics",
        "tags": [
          "operations"
        ]
      }
    },
    "/health": {
      "get": {
        "description": "Liveness only: intentionally does not touch Postgres or Redis.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "operations"
        ]
      }
    },
    "/metrics": {
      "get": {
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Metrics",
        "tags": [
          "operations"
        ]
      }
    },
    "/ready": {
      "get": {
        "operationId": "readiness_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Readiness Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readiness",
        "tags": [
          "operations"
        ]
      }
    }
  }
}